add view orders of user
