--- src/main/java/by/trubetski/QuickSolution/QuickSolutionApplication.java
@@ -1,13 +1,0 @@
-package by.trubetski.QuickSolution;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class QuickSolutionApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(QuickSolutionApplication.class, args);
-	}
-
-}
