--- src/main/java/by/trubetski/QuickSolution/exception/EmailAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package by.trubetski.QuickSolution.exception;
-
-public class EmailAlreadyExistsException extends GlobalException{
-    public EmailAlreadyExistsException(String message) {
-        super(message);
-    }
-}
