--- src/main/java/by/trubetski/QuickSolution/exception/GlobalException.java
@@ -1,7 +1,0 @@
-package by.trubetski.QuickSolution.exception;
-
-public class GlobalException extends RuntimeException{
-    public GlobalException(String message) {
-        super(message);
-    }
-}
