--- src/main/java/by/trubetski/QuickSolution/exception/UsernameAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package by.trubetski.QuickSolution.exception;
-
-public class UsernameAlreadyExistsException extends GlobalException{
-    public UsernameAlreadyExistsException(String message) {
-        super(message);
-    }
-}
