--- src/main/java/by/trubetski/QuickSolution/security/AppUsersDetails.java
@@ -1,54 +1,0 @@
-package by.trubetski.QuickSolution.security;
-
-import by.trubetski.QuickSolution.models.User;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-
-public class AppUsersDetails implements UserDetails {
-    private final User user;
-
-    public AppUsersDetails(User user) {
-        this.user = user;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return null;
-    }
-
-    @Override
-    public String getPassword() {
-        return this.user.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return this.user.getUsername();
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-
-    public User getUser(){
-        return this.user;
-    }
-}
