+++ src/main/java/by/trubetski/quickSolution/exception/EmailAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package by.trubetski.quickSolution.exception;
+
+public class EmailAlreadyExistsException extends GlobalException{
+    public EmailAlreadyExistsException(String message) {
+        super(message);
+    }
+}
