+++ src/main/java/by/trubetski/quickSolution/exception/GlobalException.java
@@ -1,0 +1,7 @@
+package by.trubetski.quickSolution.exception;
+
+public class GlobalException extends RuntimeException{
+    public GlobalException(String message) {
+        super(message);
+    }
+}
