--- src/test/java/by/trubetski/QuickSolution/QuickSolutionApplicationTests.java
@@ -1,13 +1,0 @@
-package by.trubetski.QuickSolution;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class QuickSolutionApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
