+++ src/test/java/by/trubetski/quickSolution/MyTest.java
@@ -1,0 +1,12 @@
+package by.trubetski.quickSolution;
+
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.TestPropertySource;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+@TestPropertySource(locations="classpath:test.properties")
+public class MyTest {
+}
