+++ src/test/java/by/trubetski/quickSolution/test.properties
